refactored bug about modal functionality and separate into smaller modules
