+++ app.js
@@ -1,6 +1,10 @@
-import { render, renderResult, renderKeyboard } from './view.js';
+import { render} from './view/mainView.js';
+import {renderModal} from './view/modal.js'
+import {renderKeyboard} from './view/keyboard.js'
 import { Hangman } from './hangman.js';
 import { fetchPuzzle } from './requests.js';
+
+export let game;
 
 export const startGame = async () => {
   const puzzle = await fetchPuzzle();
@@ -9,7 +13,6 @@ export const startGame = async () => {
   game.play('terror');
 };
 
-export let game;
 renderKeyboard();
 startGame();
 
@@ -19,7 +22,7 @@ keysEl.forEach((key) => {
     game.makeGuess(e.target.name); // set the puzzle
     game.setStatus(); // check status and set it
     render(); // render the main part
-    renderResult(); // render the result if it finished
+    renderModal(); // render the result if it finished
     //clearText(); // clear the input
   });
 });
