+++ index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Hangman Game</title>
     <link rel="stylesheet" href="app.css" />
-    <link rel="icon" href="./img/hangman-favi.png" type="image/x-icon" />
+    <link rel="icon" href="./img/hangman-favi.png" type="image/png" />
   </head>
   <body>
     <div class="game-container">
